+++ src/client/ions/atoms/live-painting.ts
@@ -1,0 +1,12 @@
+import { atom } from "jotai/index";
+
+import type { ImageItem } from "#/types";
+
+export const livePaintingOptionsAtom = atom({
+	brushSize: 5,
+	color: "#000000",
+});
+export const clearCounterAtom = atom(0);
+export const imageAtom = atom("");
+export const imagesAtom = atom<ImageItem[]>([]);
+export const storyImagesAtom = atom<ImageItem[]>([]);
