+++ src/client/ions/development/story.ts
@@ -1,0 +1,29 @@
+export const example = `
+# Tinkering with Time
+
+## Chapter 1: Echoes in the Garage
+
+Tommy had always been intrigued by the old garage at the end of the lane. ![inside a garage with a car](0) It was a place out of time, where the dust of decades clung to a classic blue car that never seemed to age. The shelves were lined with pots in every hue, and the walls adorned with tools that had seen the making and mending of countless machines.
+
+It was here, in the silent conversations of inanimate objects, that Tommy felt the presence of his grandfather, a mechanic of the old world whose hands were skilled in the arts of repair and restoration.
+
+## Chapter 2: Frequencies of the Past
+
+Just beyond the garage, through a door that squeaked with stories, was the room that held his grandfather’s second love: radios. ![a room with various radios](1) Rows of dials, switches, and antennas filled the space, each piece a testament to a bygone era when voices traveled through the air like invisible threads stitching together the fabric of the community.
+
+Tommy could almost hear the crackling sounds of history, the broadcasts of moon landings, and the whispers of the old radio shows that his grandfather would recount with a sparkle in his eye.
+
+## Chapter 3: The Envelope from Yesteryear
+
+The past had a way of lingering, and nowhere did it echo more profoundly than the post office at the corner of Main and Elm. ![outside of a post office](2) Its walls, painted the blue of a stormy sea, held the secrets of the town. Letters and packages passed through its doors, each carrying stories, confessions, and dreams. As Tommy entered, he noticed an envelope addressed to him in his grandfather’s familiar scrawl, postmarked from a date long past.
+
+It was as if time had curled upon itself, delivering a message meant for another age. In the quiet of the post office, with the flag outside fluttering in the wind, Tommy opened the envelope. Inside, he found the key to the garage and a note that simply read, “Fix the unfixable, tune in to the untold, and deliver the undeliverable. The journey is yours.” It was an invitation from the past, urging him to weave his own story into the tapestry of time.
+
+As he stepped out, the world seemed to hold its breath, waiting for Tommy to turn the key, not just of the garage, but of destiny. What secrets did the car hold? What stories would the radios tell? And what mysteries awaited delivery? Only time would tell, and it was time to begin.
+`;
+
+export const exampleImages = [
+	{ id: "1", dataUrl: "/demo/0.png", url: "" },
+	{ id: "2", dataUrl: "/demo/1.png", url: "" },
+	{ id: "3", dataUrl: "/demo/2.png", url: "" },
+];
