+++ src/client/ions/text-to-image/index.ts
@@ -1,0 +1,13 @@
+export const illustrationStyles = {
+	childrensBook:
+		"whimsical, watercolor, children's book illustration, soft edges, vibrant highlights, light and shadow play, dynamic perspective, expressive, anthropomorphic details, pastel background, colorful accents, imaginative scenery",
+	manga: "anime, manga art kodomo style, bright colors, simple lines, clear outlines, minimal detail, vibrant settings, fantasy elements, playful designs, friendship themes",
+	fantasyArt:
+		"detailed, imaginative, epic, vibrant color palette, mythical themes, dynamic lighting, magical elements",
+	realism:
+		"high detail, accurate proportions, lifelike textures, natural lighting, depth of field, subtle color variation, realistic expressions, meticulous backgrounds, true-to-life scenes",
+	graphicNovel:
+		"inked lines, bold shading, sequential art, visual storytelling, cinematic framing, noir tones",
+	custom: "",
+};
+export type IllustrationStyles = keyof typeof illustrationStyles;
