+++ src/shared/number.ts
@@ -1,0 +1,3 @@
+export function randomSeed() {
+	return Math.ceil(Math.random() * 1_000_000_000) + 1;
+}
